Убрал static_cast в таймере, так как он не нужен: если от uint32_t отнять uint32_t то получится тот же самый тип и результат будет вычислен по принципам беззнаковой арифметики

+++ timer.h
@@ -16,7 +16,7 @@ class Timer{
     }
 
     bool is_time(){
-      return static_cast<uint32_t>(millis() - t_start) >= dt;
+      return millis() - t_start >= dt;
     }
 
   private:
